Moved Utility/String.h to Utility/String/Transformations.h
