+++ src/Bin/OpenEnroth/OpenEnroth.cpp
@@ -22,7 +22,7 @@
 #include "Utility/Streams/FileInputStream.h"
 #include "Utility/String/Format.h"
 #include "Utility/UnicodeCrt.h"
-#include "Utility/String.h"
+#include "Utility/String/Transformations.h"
 #include "Utility/Types.h"
 
 #include "OpenEnrothOptions.h"
