+++ src/Engine/MapInfo.cpp
@@ -7,7 +7,7 @@
 #include "Utility/Memory/Blob.h"
 #include "Utility/Workaround/ToUnderlying.h"
 #include "Utility/String/Ascii.h"
-#include "Utility/String.h"
+#include "Utility/String/Transformations.h"
 
 MapStats *pMapStats;
