+++ src/Engine/Tables/AwardTable.cpp
@@ -3,7 +3,7 @@
 #include <string>
 
 #include "Utility/Memory/Blob.h"
-#include "Utility/String.h"
+#include "Utility/String/Transformations.h"
 
 std::array<Award, 105> pAwards;
