+++ src/Library/Serialization/EnumSerializer.cpp
@@ -5,7 +5,7 @@
 #include <string>
 
 #include "Utility/String/Ascii.h"
-#include "Utility/String.h"
+#include "Utility/String/Transformations.h"
 
 bool detail::EnumSerializationTable::trySerialize(uint64_t src, std::string *dst) const {
     auto pos = _stringByEnum.find(src);
