+++ src/Utility/String/Tests/Transformations_ut.cpp
@@ -3,21 +3,21 @@
 
 #include "Testing/Unit/UnitTest.h"
 
-#include "Utility/String.h"
+#include "Utility/String/Transformations.h"
 
-UNIT_TEST(String, toHexDump) {
+UNIT_TEST(StringTransformations, toHexDump) {
     EXPECT_EQ(toHexDump("1234", 2), "3132 3334");
     EXPECT_EQ(toHexDump("0000"), "30303030");
 }
 
-UNIT_TEST(String, replaceAll) {
+UNIT_TEST(StringTransformations, replaceAll) {
     EXPECT_EQ(replaceAll("123", "1", "123"), "12323");
     EXPECT_EQ(replaceAll("123", "10", "100"), "123");
     EXPECT_EQ(replaceAll("ab123ab", "ab", "zz"), "zz123zz");
     EXPECT_EQ(replaceAll("AAAA", "AA", "AAZAA"), "AAZAAAAZAA");
 }
 
-UNIT_TEST(String, split) {
+UNIT_TEST(StringTransformations, split) {
     std::vector<std::string_view> v;
 
     splitString("aa;bb;cc", ';', &v);
@@ -36,7 +36,7 @@ UNIT_TEST(String, split) {
     EXPECT_TRUE(v.empty());
 }
 
-UNIT_TEST(String, join) {
+UNIT_TEST(StringTransformations, join) {
     EXPECT_EQ(join("", '/', std::string_view("123"), std::string("321")), "/123321");
 }
